+++ extender/src/main/java/burp/tdou/fingerscan/common/Constants.java
@@ -11,7 +11,7 @@ public interface Constants {
 
     // 插件信息
     String PLUGIN_NAME = "FingerScan";
-    String PLUGIN_VERSION = "3.0.2";
+    String PLUGIN_VERSION = "3.0.3";
     boolean DEBUG = false;
 
     // 插件启动显示的信息